requirements.txt
websocket-client
supabase
pycups
requests
python-dotenv